requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
celery
redis
django-cors-headers
djangorestframework-gis
drf-yasg
python-dotenv
pytest-django
openai
google-auth
stripe
psycopg2
